Return React child instead of Component
While running this code I got the following warning in Chrome:
"Functions are not valid as a React child. This may happen if you return a Component instead of <Component /> from render. Or maybe you meant to call this function rather than return it."

and the "Sign in with Google" did not appeared as fallback.

With this small fix the warning disappeared and Sign in works fine